Приводит в порядок модули.
