+++ js/form.js
@@ -1,3 +1,9 @@
+import {createFetch} from './server-data.js';
+import {showAlert} from './utils.js';
+import {restoreParameters, createMarker} from './map.js';
+
+const FORM_SEND_ADDRESS = 'https://23.javascript.pages.academy/keksobooking';
+
 const ticketForm = document.querySelector('.ad-form');
 const filterForm = document.querySelector('.map__filters');
 const ticketFormChildren = ticketForm.querySelectorAll('fieldset');
@@ -28,5 +34,67 @@ const removeMessage = () => {
     success.remove();
   }
 };
+
+const errorMessage = document.querySelector('#error');
+const successMessage = document.querySelector('#success');
+
+const showMessage = (message) => {
+  let clickId = () => {};
+  let keydownId  = () => {};
+  document.body.appendChild(message.content);
+  document.addEventListener('click', clickId = () => {
+    removeMessage(successMessage, errorMessage);
+    document.removeEventListener('keydown', keydownId);
+  },{once:true});
+  document.addEventListener('keydown', keydownId = (evt) => {
+    if (evt.keyCode === 27) {
+      removeMessage(successMessage, errorMessage);
+      document.removeEventListener('click', clickId);
+    }
+  }, {once:true});
+};
+
+sendForm.addEventListener('submit', (evt) => {
+  evt.preventDefault();
+  const formData = new FormData(evt.target);
+  const errorMessageShow = errorMessage.cloneNode(true);
+  const successMessageShow = successMessage.cloneNode(true);
+
+  fetch(
+    FORM_SEND_ADDRESS,
+    {
+      method: 'POST',
+      body: formData,
+    },
+  ).then((response) => {
+    if (response.ok) {
+      showMessage(successMessageShow);
+      restoreParameters();
+    } else {
+      showMessage(errorMessageShow);
+    }
+  }).catch(() => {
+    showMessage(errorMessageShow);
+  });
+});
+
+const resetButton = document.querySelector('.ad-form__reset');
+
+resetButton.addEventListener('click', (evt) => {
+  evt.preventDefault();
+  restoreParameters();
+});
+
+const fetchData = createFetch(
+  (data) => {
+    data.forEach((similarObject) => {
+      createMarker(similarObject);
+    });
+  },
+  (error) => {
+    showAlert(error);
+  });
+
+fetchData();
 
 export {formEnableHandler, sendForm, removeMessage};
