+++ js/main.js
@@ -1,6 +1,6 @@
+import './map.js';
 import './form.js';
 import {similarObjects} from './map-data.js';
-import './map.js';
 import './restrictions.js';
 import './server-data.js';
 import './generation-data.js';
